# internal/collector+pkg/metrics_store: Reduce allocation & improve hot path