--- a/internal/collector/testutils.go
+++ b/internal/collector/testutils.go
@@ -31,14 +31,14 @@ type generateMetricsTestCase struct {
 	Obj         interface{}
 	MetricNames []string
 	Want        string
-	Func        func(interface{}) []metricsstore.FamilyStringer
+	Func        func(interface{}) []metricsstore.FamilyByteSlicer
 }
 
 func (testCase *generateMetricsTestCase) run() error {
 	metricFamilies := testCase.Func(testCase.Obj)
 	metricFamilyStrings := []string{}
 	for _, f := range metricFamilies {
-		metricFamilyStrings = append(metricFamilyStrings, f.String())
+		metricFamilyStrings = append(metricFamilyStrings, string(f.ByteSlice()))
 	}
 
 	metric := strings.Split(strings.Join(metricFamilyStrings, ""), "\n")