This pull request includes two performance optimizations found while testing the kube-state-metrics *v1.6.0* release candidate. Given that v1.6.0-rc.0 doesn't seem to introduce any performance regressions, these patches don't need to be included in the release.

### 1. internal/collector: Preallocate metric slice length when known upfront

Profiling kube-state-metrics shows that a lot of CPU time is spent on
`runtime.growslice`. This is related to metric generation when
converting a Kuberntes object to a slice of metrics in
`metric.FamilyGenerator`s. Oftentimes the length of the metric slice is
known upfront.

When the needed size of the metrics slice is known upfront, allocate an
appropriately sized metric slice.

### 2. pkg/metrics_store: Cache byte slices instead of strings

When converting a string to a byte slice, in Golang would do a full copy
which introduces memory allocations. On an HTTP scrape request WriteAll
in the metrics store component is called, which writes all cached
metrics into the given io.Writer. io.Writer takes a byte slice, whereas
metrics are cached as strings. The connect the two, the metrics store
component converts the metric strings to byte slices. This results in
`runtime.stringtoslicebyte` [1] being a prominent CPU user.

Instead of caching metrics as strings, cache them as byte slices,
removing the additional translation from the hot-path.

### Test setup

#### Unit benchmark

The first patch gets us some reasonable memory and runtime optimizations. Running `internal.collector.BenchmarkPodCollector` with (*new*) and without (*old*) returns the following:

```
benchmark                   old ns/op     new ns/op     delta
BenchmarkPodCollector-4     25942         24475         -5.65%

benchmark                   old allocs     new allocs     delta
BenchmarkPodCollector-4     419            395            -5.73%

benchmark                   old bytes     new bytes     delta
BenchmarkPodCollector-4     20112         19544         -2.82%
```

#### End to end test

I have compared kube-state-metrics *v1.6.0-rc.0-0* (vanilla https://github.com/kubernetes/kube-state-metrics/pull/702),  *v1.6.0-rc.0-4* (vanilla + patch 1) and *v1.6.0-rc.0-5* (vanilla + patch 1 & 2) on a 9 node cluster with AWS m5.large compute instances. During the comparison a test script would create and delete pods ranging from 250 - 5000 pods total. In addition the cluster had 8000 secrets constantly. Kube-state-metrics instances are scraped every 2s.

##### Scrape duration

Both patches seem to be neutral in regards to the Prometheus scrape duration.

![image](https://user-images.githubusercontent.com/7047859/55469099-0928ba00-5605-11e9-83c7-dd310173db90.png)


##### CPU usage

Patch 1 seems to have only a subtle impact on the CPU usage, patch 2 constantly performs better, given that we only need to convert the metric string to a byte slice once and not on every scrape.

![image](https://user-images.githubusercontent.com/7047859/55468812-7e47bf80-5604-11e9-930b-344c1ca4aa41.png)





[1] https://golang.org/src/runtime/string.go?s=4063:4115#L145