@mxinden 2019-04-05
CI is fixed. Thanks for the pointer @brancz. 

Would you mind giving this another review?